Tests for sending/receiving with INADDR_ANY
Verify that an UDP socket using INADDR_ANY can send out packets and
receive broadcasts. This behavior is important for DHCP client, which
uses INADDR_ANY to receive both broadcast and unicast. E.g. when the
client is renewing its lease, it is possible for the server to send
back DHCPNAK with either unicast or broadcast.

PiperOrigin-RevId: 300552526